+++ src/index.tsx
@@ -1,7 +1,7 @@
 import AdvancedImage from "./AdvancedImage";
-import CLDVideoPlayer from './CldVideoPlayer';
+import AdvancedVideo from './AdvancedVideo';
 export { upload, unsignedUpload, uploadBase64, rename, explicit } from "./api/upload";
 export { UploadApiOptions } from './api/upload/model/params/upload-params';
 
 export { AdvancedImage };
-export { CLDVideoPlayer };
+export { AdvancedVideo };
